+++ extensions/common-directory/src/components/action-copy-file.tsx
@@ -1,9 +1,9 @@
 import { Action, ActionPanel, Icon, showHUD } from "@raycast/api";
-import { copyFileByPath, isEmpty } from "./common-utils";
 import React from "react";
-import { DirectoryInfo } from "./directory-info";
+import { copyFileByPath, isEmpty } from "../utils/common-utils";
+import { DirectoryInfo } from "../types/directory-info";
 
-export function CopyFileActions(props: { directory: DirectoryInfo }) {
+export function ActionCopyFile(props: { directory: DirectoryInfo }) {
   const { directory } = props;
   return (
     <ActionPanel.Section>