+++ extensions/common-directory/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Common Directory Changelog
 
+## [Update directory preview] - 2022-05-27
+
+- Update directory preview
+- Add action "Open Extension Preferences"
+
 ## [Added Common Directory] - 2022-04-30
 
 - Fix Rank algorithm bugs